requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
flask-bcrypt
flask-cors
python-dotenv
pyodbc